added vote model and controller
