+++ app/controllers/votes_controller.rb
@@ -1,0 +1,2 @@
+class VotesController < ApplicationController
+end
