+++ app/helpers/votes_helper.rb
@@ -1,0 +1,2 @@
+module VotesHelper
+end
