+++ app/views/layouts/application.html.erb
@@ -19,8 +19,13 @@
           <li> <%= link_to "View all media", works_path %></li>
           <li> <%= link_to "Add new work", new_work_path %></li>
           <li> <%= link_to "View all users", works_path %></li>
-          <li> <%= link_to "Login", login_path %></li>
-    
+          <% if @current_user %>
+            <li> <%= link_to "Logged in as #{@current_user.username}", root_path %></li>
+            <li> <%= link_to "Log out", logout_path, method: :post %></li>
+          <%else%>
+            <li> <%= link_to "Login", login_path %></li>
+          <% end %>
+
         </ul>
       </nav>
     <header>
