+++ db/migrate/20190424215824_create_votes.rb
@@ -1,0 +1,9 @@
+class CreateVotes < ActiveRecord::Migration[5.2]
+  def change
+    create_table :votes do |t|
+      t.date :date
+
+      t.timestamps
+    end
+  end
+end
