+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_04_24_203503) do
+ActiveRecord::Schema.define(version: 2019_04_24_215824) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -28,6 +28,12 @@ ActiveRecord::Schema.define(version: 2019_04_24_203503) do
 
   create_table "users", force: :cascade do |t|
     t.string "username"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
+  create_table "votes", force: :cascade do |t|
+    t.date "date"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
