+++ test/controllers/votes_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+describe VotesController do
+  # it "must be a real test" do
+  #   flunk "Need real tests"
+  # end
+end
